Tidy a Corpus object from the tm package

Tidy a Corpus object from the tm package. Returns a data frame with one-row-per-document, with a text column containing the document's text, and one column for each local (per-document) metadata tag. For corpus objects from the quanteda package, see tidy.corpus.

# S3 method for Corpus
tidy(x, collapse = "\n", ...)

Arguments

x
A Corpus object, such as a VCorpus or PCorpus

collapse
A string that should be used to collapse text within each corpus (if a document has multiple lines). Give NULL to not collapse strings, in which case a corpus will end up as a list column if there are multi-line documents.

...
Extra arguments, not used

Examples

library(dplyr)   # displaying tbl_dfs

if (requireNamespace("tm", quietly = TRUE)) {
  library(tm)
  #' # tm package examples
  txt <- system.file("texts", "txt", package = "tm")
  ovid <- VCorpus(DirSource(txt, encoding = "UTF-8"),
                  readerControl = list(language = "lat"))

  ovid
  tidy(ovid)

  # choose different options for collapsing text within each
  # document
  tidy(ovid, collapse = "")$text
  tidy(ovid, collapse = NULL)$text

  # another example from Reuters articles
  reut21578 <- system.file("texts", "crude", package = "tm")
  reuters <- VCorpus(DirSource(reut21578),
                     readerControl = list(reader = readReut21578XMLasPlain))
  reuters

  tidy(reuters)
}

#> Loading required package: NLP

#> 
#> Attaching package: ‘NLP’

#> The following object is masked from ‘package:ggplot2’:
#> 
#>     annotate

#> 
#> Attaching package: ‘tm’

#> The following object is masked from ‘package:tidytext’:
#> 
#>     stopwords

#> # A tibble: 20 x 17
#>    author datetimestamp       descr… headi… id    lang… orig… topi… lewi… cgis…
#>    <chr>  <dttm>              <chr>  <chr>  <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 <NA>   1987-02-26 10:00:56 ""     DIAMO… 127   en    Reut… YES   TRAIN TRAI…
#>  2 BY TE… 1987-02-26 10:34:11 ""     OPEC … 144   en    Reut… YES   TRAIN TRAI…
#>  3 <NA>   1987-02-26 11:18:00 ""     TEXAC… 191   en    Reut… YES   TRAIN TRAI…
#>  4 <NA>   1987-02-26 11:21:01 ""     MARAT… 194   en    Reut… YES   TRAIN TRAI…
#>  5 <NA>   1987-02-26 12:00:57 ""     HOUST… 211   en    Reut… YES   TRAIN TRAI…
#>  6 <NA>   1987-02-28 20:25:46 ""     KUWAI… 236   en    Reut… YES   TRAIN TRAI…
#>  7 By Je… 1987-02-28 20:39:14 ""     INDON… 237   en    Reut… YES   TRAIN TRAI…
#>  8 <NA>   1987-02-28 22:27:27 ""     SAUDI… 242   en    Reut… YES   TRAIN TRAI…
#>  9 <NA>   1987-03-01 01:22:30 ""     QATAR… 246   en    Reut… YES   TRAIN TRAI…
#> 10 <NA>   1987-03-01 11:31:44 ""     SAUDI… 248   en    Reut… YES   TRAIN TRAI…
#> 11 <NA>   1987-03-01 18:05:49 ""     SAUDI… 273   en    Reut… YES   TRAIN TRAI…
#> 12 <NA>   1987-03-02 00:39:23 ""     GULF … 349   en    Reut… YES   TRAIN TRAI…
#> 13 <NA>   1987-03-02 00:43:22 ""     SAUDI… 352   en    Reut… YES   TRAIN TRAI…
#> 14 <NA>   1987-03-02 00:43:41 ""     KUWAI… 353   en    Reut… YES   TRAIN TRAI…
#> 15 <NA>   1987-03-02 01:25:42 ""     PHILA… 368   en    Reut… YES   TRAIN TRAI…
#> 16 <NA>   1987-03-02 04:20:05 ""     STUDY… 489   en    Reut… YES   TRAIN TRAI…
#> 17 <NA>   1987-03-02 04:28:26 ""     STUDY… 502   en    Reut… YES   TRAIN TRAI…
#> 18 <NA>   1987-03-02 05:13:46 ""     UNOCA… 543   en    Reut… YES   TRAIN TRAI…
#> 19 By BE… 1987-03-02 07:38:34 ""     NYMEX… 704   en    Reut… YES   TRAIN TRAI…
#> 20 <NA>   1987-03-02 07:49:06 ""     ARGEN… 708   en    Reut… YES   TRAIN TRAI…
#> # ... with 7 more variables: oldid <chr>, topics_cat <list>, places <list>,
#> #   people <chr>, orgs <chr>, exchanges <chr>, text <chr>