tidytext: Text Mining using 'dplyr', 'ggplot2', and Other Tidy Tools

Using tidy data principles can make many text mining tasks easier, more effective, and consistent with tools already in wide use. Much of the infrastructure needed for text mining with tidy data frames already exists in packages like 'dplyr', 'broom', 'tidyr', and 'ggplot2'. In this package, we provide functions and supporting data sets to allow conversion of text to and from tidy formats, and to switch seamlessly between tidy tools and existing text mining packages.

See also

Author

Maintainer: Julia Silge julia.silge@gmail.com (ORCID)

Authors:

David Robinson admiral.david@gmail.com

Other contributors:

Gabriela De Queiroz gabidequeiroz@gmail.com [contributor]

Colin Fay contact@colinfay.me (ORCID) [contributor]

Emil Hvitfeldt emilhhvitfeldt@gmail.com [contributor]

Os Keyes ironholds@gmail.com (ORCID) [contributor]

Kanishka Misra kmisra@purdue.edu [contributor]

Tim Mastny tim.mastny@gmail.com [contributor]

Jeff Erickson jeff@erick.so [contributor]