Tidiers for a corpus object from the quanteda package

Tidy a corpus object from the quanteda package. tidy returns a tbl_df with one-row-per-document, with a text column containing the document's text, and one column for each document-level metadata. glance returns a one-row tbl_df with corpus-level metadata, such as source and created. For Corpus objects from the tm package, see tidy.Corpus.

# S3 method for corpus
tidy(x, ...)

# S3 method for corpus
glance(x, ...)

Arguments

x
A Corpus object, such as a VCorpus or PCorpus

...
Extra arguments, not used

Details

For the most part, the tidy output is equivalent to the "documents" data frame in the corpus object, except that it is converted to a tbl_df, and texts column is renamed to text to be consistent with other uses in tidytext.

Similarly, the glance output is simply the "metadata" object, with NULL fields removed and turned into a one-row tbl_df.

Examples

if (requireNamespace("quanteda", quietly = TRUE)) {
 data("data_corpus_inaugural", package = "quanteda")

 data_corpus_inaugural

 tidy(data_corpus_inaugural)
}

#> # A tibble: 58 x 4
#>    text                                                Year President FirstName 
#>    <chr>                                              <dbl> <chr>     <chr>     
#>  1 "Fellow-Citizens of the Senate and of the House o…  1789 Washingt… George    
#>  2 "Fellow citizens, I am again called upon by the v…  1793 Washingt… George    
#>  3 "When it was first perceived, in early times, tha…  1797 Adams     John      
#>  4 "Friends and Fellow Citizens:\n\nCalled upon to u…  1801 Jefferson Thomas    
#>  5 "Proceeding, fellow citizens, to that qualificati…  1805 Jefferson Thomas    
#>  6 "Unwilling to depart from examples of the most re…  1809 Madison   James     
#>  7 "About to add the solemnity of an oath to the obl…  1813 Madison   James     
#>  8 "I should be destitute of feeling if I was not de…  1817 Monroe    James     
#>  9 "Fellow citizens, I shall not attempt to describe…  1821 Monroe    James     
#> 10 "In compliance with an usage coeval with the exis…  1825 Adams     John Quin…
#> # … with 48 more rows