get_sentiments {tidytext}	R Documentation

Get a tidy data frame of a single sentiment lexicon

Description

Get specific sentiment lexicons in a tidy format, with one row per word, in a form that can be joined with a one-word-per-row dataset. Each of these comes from the included sentiments data frame, but this performs the filtering for a specific lexicon, and removes columns that are not used in that lexicon.

Usage

get_sentiments(lexicon = c("afinn", "bing", "nrc", "loughran"))

Arguments

lexicon
The sentiment lexicon to retrieve; either "afinn", "bing", "nrc", or "loughran"

Value

A tbl_df with a word column, and either a sentiment column (if lexicon is not "afinn") or a numeric score column (if lexicon is "afinn").

Examples

library(dplyr)
get_sentiments("afinn")
get_sentiments("bing")

[Package tidytext version 0.1.8 Index]