get_stopwords {tidytext}	R Documentation

Get a tidy data frame of a single stopword lexicon

Description

Get a specific stop word lexicon via the stopwords packages's stopwords function, in a tidy format with one word per row.

Usage

get_stopwords(language = "en", source = "snowball")

Arguments

language
The language of the stopword lexicon specified as a two-letter ISO code, such as "es", "de", or "fr". Default is "en" for English. Use stopwords_getlanguages from stopwords to see available languages.

source
The source of the stopword lexicon specified. Default is "snowball". Use stopwords_getsources from stopwords to see available sources.

Value

A tibble with two columns, word and lexicon. The parameter lexicon is "quanteda" in this case.

Examples

library(dplyr)
get_stopwords()
get_stopwords(source = "smart")
get_stopwords("es", "snowball")
get_stopwords("ru", "snowball")

[Package tidytext version 0.1.6 Index]