parts_of_speech {tidytext}	R Documentation

Parts of speech for English words from the Moby Project

Description

Parts of speech for English words from the Moby Project by Grady Ward. Words with non-ASCII characters and items with a space have been removed.

Usage

parts_of_speech

Format

A data frame with 205,985 rows and 2 variables:

word

An English word

pos

The part of speech of the word. One of 13 options, such as "Noun", "Adverb", "Adjective"

Source

https://archive.org/details/mobypartofspeech03203gut

Examples

library(dplyr)

parts_of_speech

parts_of_speech %>%
  count(pos, sort = TRUE)

[Package tidytext version 0.1.6 Index]